-Added appointment system. Surgery model,service and repository(+impl). -Changed bootstrap styles and added theme.css. -Sections added for homepage and navbar. -All admin jsp files updates. -Application Properties changed. -Some DTO classes changed.

+++ src/main/java/com/vetapp/veterinarysystem/model/Pet.java
@@ -16,7 +16,7 @@ public class Pet {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "PetID")
-    private int petID;
+    private Long petID;
 
     @Column(name = "Name", nullable = false, length = 50)
     private String name;
